+++ packages/builder/package.json
@@ -15,12 +15,10 @@
     "npm": ">=3.0.0"
   },
   "dependencies": {
-    "react": "^16.8.3",
-    "react-dom": "^16.7.0",
     "moment": "^2.24.0",
-    "@uform/react": "^0.0.9",
-    "@uform/utils": "^0.0.9",
-    "@uform/validator": "^0.0.9",
+    "@uform/react": "^0.1.0-aplha.3",
+    "@uform/utils": "^0.1.0-aplha.3",
+    "@uform/validator": "^0.1.0-aplha.3",
     "pascal-case": "^2.0.1",
     "react-eva": "^1.0.0",
     "classnames": "^2.2.5",
