+++ packages/builder/src/components/fields/index.js
@@ -13,6 +13,7 @@ import {
 } from '../../actions'
 import uniqBy from 'lodash.uniqby'
 import styled from 'styled-components'
+import { indexStyle } from './style'
 
 class FieldList extends Component {
   static propTypes = {
@@ -126,47 +127,7 @@ const mapDispatchToProps = dispatch => ({
 })
 
 const StyledFieldList = styled(FieldList)`
-  .field-list {
-    font-size: 0;
-    li {
-      overflow: hidden;
-      margin-bottom: 12px;
-      padding: 0 8px;
-      display: inline-block;
-      width: 33.33%;
-      height: 75px;
-      font-size: 12px;
-      text-align: center;
-      color: ${props => props.theme.whiteColor};
-      box-sizing: border-box;
-      transition: all 0.1s ease;
-      cursor: pointer;
-      &:hover {
-        background: ${props => props.theme.compHoverBgColor};
-      }
-      span {
-        display: block;
-        margin: auto;
-        max-width: 36px;
-        height: 32px;
-        word-break: break-all;
-      }
-    }
-  }
-  .field-icon {
-    display: block;
-    margin: 0 auto 12px;
-    height: 30px;
-    background-repeat: no-repeat;
-    background-position: 50% 50%;
-  }
-  @media screen and (max-width: 834px) {
-    .field-list {
-      li {
-        width: 100%;
-      }
-    }
-  }
+  ${indexStyle}
 `
 
 class StyledFieldListComp extends React.Component {
